refactor(shared): cleaned up dto's
