+++ shared/dtos/requests/book-trip-request.dto.ts
@@ -1,9 +1,9 @@
 import { Type } from "class-transformer";
 import { ArrayNotEmpty, IsNotEmpty, ValidateNested } from "class-validator";
-import { Customer } from "../models/customer";
-import { Seat } from "../models/seat";
+import { Customer } from "../../models/customer";
+import { Seat } from "../../models/seat";
 
-export class BookTripDto {
+export class BookTripRequestDto {
   @IsNotEmpty()
   @ValidateNested()
   @Type(() => Customer)
