+++ shared/dtos/responses/trip-search-result.dto.ts
@@ -1,5 +1,5 @@
-import { TripPoint } from "../models/trip-point";
-import { TripResponse } from "../models/trip-response";
+import { TripPoint } from "../../models/trip-point";
+import { TripResponse } from "../../models/trip-response";
 
 export class BookingResultDto {
   // Trip data including what switches to make and what trains to travel with
